Add rating field to the ad

+++ app/src/main/java/com/codebusters/idealizeprojectdraft/AddingAdsActivity.kt
@@ -209,7 +209,7 @@ class AddingAdsActivity : AppCompatActivity() {
             uri!!,
             binding.autoCompleteTextViewSellScreen.text.toString(),
             myTags.adVisible,
-            uid+"_"+(idealizeUser.adCount+1),
+            uid+"_"+(idealizeUser.adCount.toInt()+1),
             uid
             )
         return getDateTime(item)
